cmt 37 [16/07]
Se agrega logica de tran_vta_lote para las anulaciones, y tener registro
de el stock en las ventas contra los lotes. Se arregla problemas de
actualizacion en pers fisica y juridica, ya que con el update on
cascade, no se estaba haciendo bien la logica luego de actualizar la
persona (se intentaba actualizar tambien la pk de pf o pj).
Se cambia tipo de dato de nroTransac a Integer (como especifica la
base).

+++ GestPedDsk/src/gpd/interfaces/transaccion/IPersTranLinea.java
@@ -8,9 +8,9 @@ import gpd.exceptions.PersistenciaException;
 
 public interface IPersTranLinea {
 
-	public TranLinea obtenerTranLineaPorId(Long nroTransac, Integer idProducto) throws PersistenciaException;
+	public TranLinea obtenerTranLineaPorId(Integer nroTransac, Integer idProducto) throws PersistenciaException;
 	public List<TranLinea> obtenerListaTranLinea(Transaccion transac) throws PersistenciaException;
 	public Integer guardarListaTranLinea(List<TranLinea> tranLinea) throws PersistenciaException;
-	public Integer eliminarTranLinea(Long nroTransac) throws PersistenciaException;
+	public Integer eliminarTranLinea(Integer nroTransac) throws PersistenciaException;
 	
 }
