+++ GestPedDsk/src/gpd/persistencia/conector/Conector.java
@@ -292,15 +292,15 @@ public abstract class Conector {
 	}
 	
 	
-	public static Long obtenerSecuencia(String nombreSec) throws ConectorException {
-		Long resultado = null;
+	public static Integer obtenerSecuencia(String nombreSec) throws ConectorException {
+		Integer resultado = null;
 		StringBuilder sb = new StringBuilder();
 		sb.append("select nextval('").append(nombreSec).append("') as seq");
 		GenSqlSelectType genType = new GenSqlSelectType(sb.toString());
 		ResultSet rs = selectGeneric(genType);
 		try {
 			if(rs.next()) {
-				resultado = rs.getLong("seq");
+				resultado = rs.getInt("seq");
 			}
 		} catch (SQLException e) {
 			logger.error(e.getMessage(), e);
